pyproject.toml
[tool.poetry]
name = "ddd"
version = "0.1.0"
description = ""
authors = ["Frank Chacon <frchacon@mercadona.es>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
asgiref = "^3.7.0"
django = "^4.1.0"
django-extensions = "^3.2.1"
django-filter = "^23.2"
djangorestframework = "^3.14.0"
djangorestframework-jsonapi = "^6.0.0"
inflection = "^0.5.1"
python-dotenv = "^1.0.0"
pytz = "^2023.3"
sqlparse = "^0.4.4"
tzdata = "^2023.3"
psycopg2-binary = "^2.9.6"


[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
pytest-django = "^4.5.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "ddd.settings"
pythonpath = "ddd"
